refactor: move time function out of root.c and into time.c
